requirements.txt
openai
python-dotenv
tqdm
scikit-learn
numpy
